==> Code/04-data-preprocessing-news-companies.ipynb <==
# Data Preprocessing
> Author: Chandan Rao (chandankuma4@iisc.ac.in)

import json
import torch
import pandas as pd
from datetime import datetime
from transformers import pipeline
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

# Load the dataset
df = pd.read_csv('../Dataset/newsapi/raw/NewsDataset.csv')

# Extracting the name value from the source column
df['source'] = df['source'].apply(lambda x: eval(x)["name"] if isinstance(x, str) else x)

df['source']

# Converting 'publishedAt' to datetime and extracting the date and time
df['publishedAt'] = pd.to_datetime(df['publishedAt']).dt.tz_convert(None)

# Removing microseconds from the datetime
df['publishedAt'] = df['publishedAt'].dt.floor('s')

df.publishedAt.head()

# Remove duplicate rows based on all columns
df.drop_duplicates(inplace=True)

# Retain only the source, author, title, description, publishedAt, url, and content columns
df_filtered = df[['source','author', 'title', 'description', 'publishedAt', 'url', 'content','search_query']]
df_filtered.head()

# Rename the search_query field to companyName and author to authors
df_filtered.rename(columns={'search_query': 'companyName','author':'authors', 'content':'summary'}, inplace=True)

df_filtered.head()

# Remove the last two words from the companyName field values
df_filtered.loc[:, 'companyName'] = df_filtered['companyName'].apply(lambda x: ' '.join(x.split()[:-2]))

df_filtered.head()

# Get base url

from urllib.parse import urlparse

def get_base_url(url):
    parsed_url = urlparse(url)
    base_url = f"{parsed_url.scheme}://{parsed_url.netloc}"
    return base_url

# Applying the function to the 'url' column
df_filtered['source_url'] = df_filtered['url'].apply(get_base_url)

df_filtered.head()

df_filtered.head()

# Remove all rows that contain any null value

# Display the count of null values
print(df_filtered.isnull().sum())

# Remove all rows with any null values
df_cleaned = df_filtered.dropna()

# Display the cleaned DataFrame
print(df_cleaned)

# Verify the count of null values
df_cleaned.isnull().sum()

# function computes the compound sentiment score for a given piece of text using the VADER sentiment analysis tool,
# which is useful for quickly assessing whether the sentiment of the text is positive, negative, or neutral.

def get_vader_score(text):
    analyzer = SentimentIntensityAnalyzer()
    sentiment = analyzer.polarity_scores(text)
    return sentiment["compound"]

# Add a new column vader to the DataFrame, which contains the sentiment scores generated by the get_vader_score function.
# The function presumably calculates sentiment analysis using the VADER (Valence Aware Dictionary and sEntiment Reasoner) method,
# which provides a score indicating the sentiment of the text in the content column

df_cleaned['summary_vader'] = df_cleaned['summary'].apply(get_vader_score)

df_cleaned['description_vader'] = df_cleaned['description'].apply(get_vader_score)

df_cleaned.head()

df_cleaned.shape

# Save the pre-processed file
df_cleaned.to_csv('../Dataset/newsapi/preprocessed/company_news_preprocessed.csv', index=False)
