==> Code/06-data-preprocessing-stocks.ipynb <==
<!-- Data Preprocessing -->

<!-- Author : Danish Akhlaq
Email: danishakhlaq@iisc.ac.in -->

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn import tree
from sklearn.tree import DecisionTreeClassifier
import plotly.graph_objects as go

# loading the  data 
df_nifty_2 = pd.read_csv('C:/Users/609370801/DSP_project/Dataset/Nifty/prepossed/NIFTY 50-26-10-2023-to-25-10-2024.csv')
df_nifty_2.head()

# Renaming the sentiment column as 'Daily Change'
df_nifty_2.rename(columns={'Sentiments': 'Daily_Change'}, inplace=True)
df_nifty_2.head()


# # Data cleaning and pre-processing

# Checking for Null values
df_nifty_2.isnull().sum()

# Checking for intended spaces
df_nifty_2.columns = df_nifty_2.columns.str.strip()

# Converting Date to date time format
try:
    df_nifty_2['Date'] = pd.to_datetime(df_nifty_2['Date'])
except KeyError:
    print("Column 'Date' does not exist in the DataFrame.")



# Creating a new column "Daily_change_percent to calculate the %change in nifty points on a daily basis
df_nifty_2['Daily_change_percent'] = (df_nifty_2['Daily_Change'] / df_nifty_2['Open']) * 100
df_nifty_2.head()

# downloading processed data set to csv
df_nifty_2.to_csv('NIFTY 50-26-10-2023-to-25-10-2024_preprocessing.csv', index=False, encoding='utf-8')
