==> Code/11-market-news-sentiment.ipynb <==
import pandas as pd
import numpy as np
from transformers import BertModel, BertTokenizer
import matplotlib.pyplot as plt
from tqdm.notebook import tqdm

import torch
import torch.nn as nn
import torch.optim as optim
import torch.nn.functional as F
from torch.utils.data import Dataset, DataLoader, random_split

torch.cuda.is_available()
torch.cuda.empty_cache()

# CONSTANTS
DEVICE = torch.device("cuda" if torch.cuda.is_available() else "cpu")
BERT_MODEL = 'google-bert/bert-base-uncased'
BATCH_SIZE = 16

class BERTSentimentClassifier(nn.Module):
    def __init__(self, num_classes=3):
        super().__init__()
        # Load pre-trained BERT model
        self.bert = BertModel.from_pretrained(BERT_MODEL).train().to(DEVICE)
        # Sequential block for dense layers
        self.classifier = nn.Sequential(
            nn.Dropout(p=0.3),
            nn.Linear(self.bert.config.hidden_size, 256),
            nn.ReLU(),
            nn.Dropout(p=0.3),
            nn.Linear(256, num_classes)
        ).train().to(DEVICE)
        # Initialize weights
        BERTSentimentClassifier.initialize_weights(self.classifier)


    def forward(self, inputs):
        embeddings = self.bert(**inputs).pooler_output
        logits = self.classifier(embeddings)
        return logits


    @staticmethod
    def initialize_weights(model):
        with torch.no_grad():
            for m in model.modules():
                if isinstance(m, (nn.Conv2d, nn.ConvTranspose2d, nn.BatchNorm2d, nn.Linear)):
                    nn.init.normal_(m.weight, 0.0, 0.02)

class NewsDataset(Dataset):
    def __init__(self, dataframe, max_length=512):
        self.tokenizer = BertTokenizer.from_pretrained(BERT_MODEL)
        self.max_length = max_length
        self.data = dataframe

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):
        summary_inputs = self.tokenizer(
            str(self.data['summary'][index]),
            max_length=self.max_length,
            padding='max_length',
            return_tensors='pt',
            truncation=True,
        )
        description_inputs = self.tokenizer(
            str(self.data['description'][index]),
            max_length=self.max_length,
            padding='max_length',
            return_tensors='pt',
            truncation=True,
        )

        summary_dict = {
            'input_ids': summary_inputs['input_ids'].squeeze(),
            'token_type_ids': summary_inputs['token_type_ids'].squeeze(),
            'attention_mask': summary_inputs['attention_mask'].squeeze(),
        }

        description_dict = {
            'input_ids': description_inputs['input_ids'].squeeze(),
            'token_type_ids': description_inputs['token_type_ids'].squeeze(),
            'attention_mask': description_inputs['attention_mask'].squeeze(),
        }

        return index, summary_dict, description_dict

    @staticmethod
    def collate_fn(batch):
        indices = []

        summary_input_ids = []
        summary_token_type_ids = []
        summary_attention_mask = []

        description_input_ids = []
        description_token_type_ids = []
        description_attention_mask = []

        for item in batch:
            indices.append(item[0])
            summary_input_ids.append(item[1]['input_ids'])
            summary_token_type_ids.append(item[1]['token_type_ids'])
            summary_attention_mask.append(item[1]['attention_mask'])

            description_input_ids.append(item[2]['input_ids'])
            description_token_type_ids.append(item[2]['token_type_ids'])
            description_attention_mask.append(item[2]['attention_mask'])

        summary_dict = {
            'input_ids': torch.stack(summary_input_ids),
            'token_type_ids': torch.stack(summary_token_type_ids),
            'attention_mask': torch.stack(summary_attention_mask),
        }

        description_dict = {
            'input_ids': torch.stack(description_input_ids),
            'token_type_ids': torch.stack(description_token_type_ids),
            'attention_mask': torch.stack(description_attention_mask),
        }

        return indices, summary_dict, description_dict


model = BERTSentimentClassifier()

state_dict = torch.load('../Models/sentiment_classification/bert_classifier.model', map_location=DEVICE)
# Remove the 'module.' prefix from the keys if present.
new_state_dict = {}
for k, v in state_dict.items():
    if k.startswith('module.'):
        new_state_dict[k[len('module.'):]] = v
    else:
        new_state_dict[k] = v
# Now load the modified state dictionary into your model.
model.load_state_dict(new_state_dict, strict=False)

model.eval()

news_df = pd.read_excel('market_article_df.xlsx')
news_df.head()

dataset = NewsDataset(news_df)
loader = DataLoader(dataset, batch_size=BATCH_SIZE, shuffle=False, collate_fn=NewsDataset.collate_fn)

for batch in (pbar := tqdm(loader)):
    torch.cuda.empty_cache()
    indices, summary, description = batch
    summary = {k: v.to(DEVICE) for k, v in summary.items()}
    description = {k: v.to(DEVICE) for k, v in description.items()}
    with torch.amp.autocast('cuda'):
        summary_logits = model(summary).cpu()
        description_logits = model(description).cpu()

    for idx, index in enumerate(indices):
        news_df.loc[index, 'summary_sentiment'] = summary_logits[idx].argmax().item()
        news_df.loc[index, 'description_sentiment'] = description_logits[idx].argmax().item()

news_df.head()

news_df.to_excel('../Dataset/news_ratings/market_article_df_with_ratings.xlsx', index=False)
