==> Code/13-news-ratings-aggregation.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt

market_news_df = pd.read_excel('../Dataset/news_ratings/market_article_df_with_ratings.xlsx', parse_dates=['datetime'])
market_news_df.head()

company_news_df = pd.read_excel('../Dataset/news_ratings/company_article_df_with_ratings.xlsx', parse_dates=['publishedAt'])
company_news_df.head()

set(market_news_df.columns)-set(company_news_df.columns)

set(company_news_df.columns)-set(market_news_df.columns)

market_news_df['market_cap_ratio'] = 1
market_news_df.rename(columns={'datetime': 'publishedAt'}, inplace=True)
market_news_df[sorted(market_news_df.columns)].head()

company_news_df.drop(columns=['Unnamed: 0', 'companyName', 'companyTicker', 'source'], inplace=True)
company_news_df[sorted(company_news_df.columns)].head()

news_df = pd.concat([market_news_df, company_news_df], ignore_index=True)
news_df.sample(5)

def adjust_to_market_day(dt):
    # Shift to the previous day if before 9:30 AM
    if dt.time() < pd.Timestamp("09:30:00").time():
        return dt - pd.Timedelta(days=1)
    return dt

news_df['market_date'] = news_df['publishedAt'].apply(adjust_to_market_day).dt.date
news_df['summary_sentiment'] = news_df['summary_sentiment'].apply(lambda x: x - 1)
news_df['description_sentiment'] = news_df['description_sentiment'].apply(lambda x: x - 1)

def weighted_average(df, value_col, weight_col):
    return (df[value_col] * df[weight_col]).sum() / df[weight_col].sum()

cols = ['summary_vader', 'summary_sentiment', 'description_vader', 'description_sentiment']
grp = news_df.groupby('market_date')
agg_df = grp.apply(
    lambda x: pd.Series(
        {'news_count': x.shape[0]} | 
        {col: weighted_average(x, col, 'market_cap_ratio') for col in cols}
    )
).reset_index()
agg_df.tail()

agg_df.to_csv('../Dataset/aggregated/aggregated_news_data.csv', index=False)
