==> Code/15-stocks_news_sentiments_correlation_study.ipynb <==
# Importing  common python libraries 
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn import tree
from sklearn.tree import DecisionTreeClassifier
import plotly.graph_objects as go
import warnings
warnings.filterwarnings('ignore')

# Userdefined functions


# Function to load the csv files in data frame

def load_csv_to_dataframe(file_path):
    """
    Load a CSV file into a pandas DataFrame.

    Parameters:
    file_path (str): The path to the CSV file.

    Returns:
    pd.DataFrame: The loaded DataFrame.
    """
    try:
        df = pd.read_csv(file_path)
        print("CSV file loaded successfully.")
        return df
    except Exception as e:
        print(f"An error occurred while loading the CSV file: {e}")
        return None


# *************************************************************************


# ################# Function for creating features########333

def create_features(df):
    df = df.copy().sort_index()

    df['Close_rolling_mean_7'] = df['Close'].rolling(window=7).mean()
    df['Close_rolling_std_7'] = df['Close'].rolling(window=7).std()

    # Exponential Moving Average (EMA)
    df['Close_ema_7'] = df['Close'].ewm(span=7, adjust=False).mean()

    # Volume-based Features
    df['Volume_rolling_mean_7'] = df['Shares Traded'].rolling(window=7).mean()
    df['Volume_pct_change'] = df['Shares Traded'].pct_change()

    # VADER & Sentiment Features
    df['summary_vader_rolling_mean_7'] = df['summary_vader'].rolling(window=7).mean()
    df['summary_sentiment_rolling_mean_7'] = df['summary_sentiment'].rolling(window=7).mean()
    df['description_vader_rolling_mean_7'] = df['description_vader'].rolling(window=7).mean()
    df['description_sentiment_rolling_mean_7'] = df['description_sentiment'].rolling(window=7).mean()

    # Drop rows with NaN values created by lagging (optional)
    df = df.dropna()

    return df

##################



# loading data into DF

df_stocks = load_csv_to_dataframe('C:\/Users\/609370801\/DSP_project\/DSP_project\/Dataset\/stocks\/preprocessed\/NIFTY 50-26-10-2023-to-25-10-2024_preprocessing.csv')
df_news_sentiments = load_csv_to_dataframe('C:\/Users\/609370801\/DSP_project\/DSP_project\/Dataset\/aggregated\/aggregated_news_data.csv')
# df_stocks.head()
# df_news_sentiments.head()


df_news_sentiments.head()

df_stocks.head()



# Function to merge stock data with average sentiment on the basis of date

df_news_sentiments['market_date'] = pd.to_datetime(df_news_sentiments['market_date'])
df_stocks['Date'] = pd.to_datetime(df_stocks['Date'])
work_correlation_df = pd.merge(df_news_sentiments, df_stocks, left_on='market_date', right_on='Date', how='inner')
work_correlation_df.head()



work_correlation_df.info()

# Perform correlation study 
# correlation = correlation_study(work_correlation_df, 'Open', 'summary_sentiment')
correlation = work_correlation_df.corr()
# Display correlation results
print("\nCorrelation Matrix:")
correlation
# print(correlation)

<!-- Strong Positive Correlations:

Open, High, Low, and Close: These variables are highly correlated with each other, with correlation coefficients around 0.98 to 0.99. This indicates that when one of these variables increases, the others tend to increase as well.
Shares Traded and Turnover (₹ Cr): There is a strong positive correlation (r = 0.95), suggesting that higher trading volumes are associated with higher turnover.
Daily Change and Daily Change Percent: These two variables are almost perfectly correlated (r = 0.99), which is expected as they represent similar concepts.
Moderate Positive Correlations:

Sentiment Scores: There are moderate positive correlations between the sentiment scores derived from summaries and descriptions:
Summary Vader and Summary Sentiment (r = 0.74)
Summary Vader and Description Vader (r = 0.64)
Summary Vader and Description Sentiment (r = 0.57)
Summary Sentiment and Description Vader (r = 0.70)
Summary Sentiment and Description Sentiment (r = 0.78)
Description Vader and Description Sentiment (r = 0.70)
Date with Open, High, Low, and Close: The date shows moderate positive correlations with these variables (r ≈ 0.87 to 0.89), indicating a general upward trend over time.
News Count with Market Variables: News count has moderate positive correlations with Date (r = 0.41), Open (r = 0.45), High (r = 0.48), Low (r = 0.43), and Close (r = 0.46).
Weak Negative Correlations:

Market Date with Sentiment Scores: There are weak negative correlations between market date and sentiment scores, such as summary sentiment (r = -0.23) and description sentiment (r = -0.18).
These findings suggest that market variables like Open, High, Low, and Close are closely related, and sentiment scores from summaries and descriptions are moderately correlated. Additionally, trading volume and turnover are strongly linked, and there is a general upward trend in market variables over time. News count also shows some correlation wi -->th market activity.
